pyproject.toml
[tool.poetry]
name = "github-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.8"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
